Uncomment tests
