+++ tests/fXmlRpc/Serializer/XmlWriterSerializerTest.php
@@ -25,12 +25,33 @@
 namespace fXmlRpc\Serializer;
 
 use fXmlRpc\ExtensionSupportInterface;
+use fXmlRpc\Value\Base64;
 
 class XmlWriterSerializerTest extends AbstractSerializerTest
 {
     protected function setUp(): void
     {
         $this->serializer = new XmlWriterSerializer();
+    }
+
+    public function provideTypes()
+    {
+        return array(
+            array('string', 'test string', 'test string'),
+            array('int', 2, '2'),
+            array('int', -2, '-2'),
+            array('double', 1.2, '1.2'),
+            array('double', -1.2, '-1.2'),
+            array('boolean', true, '1'),
+            array('boolean', false, '0'),
+            array(
+                'dateTime.iso8601',
+                \DateTime::createFromFormat('Y-m-d H:i:s', '1998-07-17 14:08:55', new \DateTimeZone('UTC')),
+                '19980717T14:08:55'
+            ),
+            array('base64', Base64::serialize('string'), "c3RyaW5n\n"),
+            array('string', 'Ümläuts', '&#220;ml&#228;uts'),
+        );
     }
 
     public function testDisableNilExtension()
